Give ClipParameters their own files
