+++ src/tracks/playabletrack/wavetrack/ui/WaveformView.cpp
@@ -11,10 +11,11 @@ Paul Licameli split from WaveChannelView.cpp
 
 #include "WaveformView.h"
 
-#include "WaveformCache.h"
-#include "WaveformVRulerControls.h"
+#include "ClipParameters.h"
 #include "WaveChannelView.h"
 #include "WaveChannelViewConstants.h"
+#include "WaveformCache.h"
+#include "WaveformVRulerControls.h"
 
 #include "SampleHandle.h"
 #include "../../../ui/EnvelopeHandle.h"
